+++ core/src/main/java/arez/CircularBuffer.java
@@ -8,7 +8,6 @@ import javax.annotation.Nullable;
 /**
  * A circular buffer implementation.
  */
-@SuppressWarnings( "Duplicates" )
 final class CircularBuffer<T>
 {
   /**
@@ -76,18 +75,27 @@ final class CircularBuffer<T>
     insertLast( Objects.requireNonNull( object ) );
   }
 
-  void addFirst( @Nonnull final T object )
-  {
-    tryGrowBeforeAdd();
-    insertFirst( Objects.requireNonNull( object ) );
-  }
-
   private void tryGrowBeforeAdd()
   {
-    final int newSize = size() + 1;
-    if ( newSize > _elements.length )
+    final int currentSize = size();
+    if ( currentSize + 1 > _elements.length )
     {
-      grow( newSize );
+      final int newSize = ( ( _elements.length - 1 ) * 2 ) + 1;
+      @SuppressWarnings( "unchecked" )
+      final T[] elements = (T[]) new Object[ newSize ];
+      int j = 0;
+      for ( int i = 0; i < currentSize; i++ )
+      {
+        final int index = ( _head + i ) % _elements.length;
+        elements[ j ] = _elements[ index ];
+        _elements[ index ] = null;
+        j++;
+      }
+
+      _elements = elements;
+      _head = 0;
+      _tail = j;
+      _isWrappedBuffer = false;
     }
   }
 
@@ -130,33 +138,6 @@ final class CircularBuffer<T>
     return 0 == size();
   }
 
-  void truncate( final int size )
-  {
-    if ( _elements.length > size )
-    {
-      shrink( size );
-    }
-  }
-
-  /**
-   * Insert specified object at head of the buffer.
-   *
-   * @param object the object.
-   */
-  private void insertFirst( @Nonnull final T object )
-  {
-    if ( 0 == _head )
-    {
-      _head = _elements.length - 1;
-      _isWrappedBuffer = true;
-    }
-    else
-    {
-      _head--;
-    }
-    _elements[ _head ] = object;
-  }
-
   /**
    * Insert specified object at tail of the buffer.
    *
@@ -171,60 +152,6 @@ final class CircularBuffer<T>
       _tail = 0;
       _isWrappedBuffer = true;
     }
-  }
-
-  /**
-   * Increase the size of the buffer.
-   */
-  private void grow( final int minCapacity )
-  {
-    int newSize = _elements.length;
-    while ( newSize < minCapacity )
-    {
-      newSize = ( ( newSize - 1 ) * 2 ) + 1;
-    }
-
-    @SuppressWarnings( "unchecked" )
-    final T[] elements = (T[]) new Object[ newSize ];
-
-    int j = 0;
-    final int size = size();
-    for ( int i = 0; i < size; i++ )
-    {
-      final int index = ( _head + i ) % _elements.length;
-      elements[ j ] = _elements[ index ];
-      _elements[ index ] = null;
-      j++;
-    }
-
-    _elements = elements;
-    _head = 0;
-    _tail = j;
-    _isWrappedBuffer = false;
-  }
-
-  /**
-   * Decrease the size of the buffer.
-   */
-  private void shrink( final int maxCapacity )
-  {
-    @SuppressWarnings( "unchecked" )
-    final T[] elements = (T[]) new Object[ maxCapacity ];
-
-    int j = 0;
-    final int size = Math.min( size(), maxCapacity );
-    for ( int i = 0; i < size; i++ )
-    {
-      final int index = ( _head + i ) % _elements.length;
-      elements[ j ] = _elements[ index ];
-      _elements[ index ] = null;
-      j++;
-    }
-
-    _elements = elements;
-    _head = 0;
-    _tail = j;
-    _isWrappedBuffer = false;
   }
 
   /*