+++ src/test/java/arez/dom/AbstractTest.java
@@ -1,7 +1,6 @@
 package arez.dom;
 
 import arez.ArezTestUtil;
-import org.realityforge.braincheck.BrainCheckTestUtil;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 
@@ -11,14 +10,12 @@ public abstract class AbstractTest
   @BeforeMethod
   protected void beforeTest()
   {
-    BrainCheckTestUtil.resetConfig( false );
     ArezTestUtil.resetConfig( false );
   }
 
   @AfterMethod
   protected void afterTest()
   {
-    BrainCheckTestUtil.resetConfig( true );
     ArezTestUtil.resetConfig( true );
   }
 }