+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v0.15](https://github.com/arez/arez-dom/tree/v0.15) (2019-03-06)
+[Full Changelog](https://github.com/arez/arez-dom/compare/v0.14...v0.15)
 
 * Merge the [arez-browserlocation](https://github.com/arez/arez-browserlocation) project into `arez-dom`.
 * Merge the [arez-networkstatus](https://github.com/arez/arez-networkstatus) project,